feat: Adjust scripts to not use Swarm and work with Podman
Generally, the scripts behave the same as before with comparable logic,
but in order for the scripts to work in a non-Swarm environment and to
be cross-compatible with both Docker and Podman, we needed to make a
number of adjustments/fixes that were applied several times among the
many scripts:

- Docker and Podman aren't consistent in how they name the individual
  containers and volumes. But they do consistently use labels, and
  `com.docker.compose.project` will always hold the name of the separate
  environments (medce/pharmce). So we filter containers that way, and
  then further filter for the desired label or container by just using
  `grep` on the results.

- The `podman-compose` project behaves differently than `docker-compose`
  when handling variables. You have to be sure and export the variables
  it needs to see when substituting variables in the docker-compose.yml
  file. So, we source `export-env.sh` which both handles this logic and
  eliminates some of the repetition from the scripts.

- The `calc-docker-compose-path.sh` script is renamed and set to handle
  not just which version to use, but it provides the full path of the
  correct docker-compose.yml file.

- When calling `docker-compose`, we need to `cd` into the main project
  directory (no longer the environment directory). This is necessary
  because `podman-compose` doesn't seem to figure out what directory the
  docker-compose.yml file is in, and instead may reference paths from
  where the script was executed from, which could be anywhere. So we
  reset things by explicitly `cd`ing into the directory in a subshell.

- Lastly, when using Podman, we want to put the containers in a pod
  since that facilitates better systemd service creation. So, we have
  logic to handle pods in the start/stop scripts.

+++ update-config.sh
@@ -25,6 +25,9 @@ for mname in $mnames; do
       done
    fi
 
+   # If config.php does not exist, skip the rest of this
+   [ ! -f "$src_path/config.php" ] && continue
+
    # Replace values in config
    sed -i -e "/'dbport'/ s/> .*,/> 3306,/" "$src_path/config.php"
    sed -i -e "/CFG->dbhost/ s/'.*'/'mariadb'/" "$src_path/config.php"
